Refactors block editor import conversion into a separate file.
